+++ db/db_impl.cc
@@ -201,8 +201,8 @@ DBImpl::DBImpl(const DBOptions& options, const std::string& dbname)
                              immutable_db_options_.table_cache_numshardbits);
 
   versions_.reset(new VersionSet(dbname_, &immutable_db_options_, env_options_,
-                                 table_cache_.get(), write_buffer_manager_,
-                                 &write_controller_));
+                                 this, table_cache_.get(),
+                                 write_buffer_manager_, &write_controller_));
   column_family_memtables_.reset(
       new ColumnFamilyMemTablesImpl(versions_->GetColumnFamilySet()));