+++ db/db_wal_test.cc
@@ -701,8 +701,8 @@ class RecoveryTestHelper {
     WriteController write_controller;
 
     versions.reset(new VersionSet(test->dbname_, &db_options, env_options,
-                                  table_cache.get(), &write_buffer_manager,
-                                  &write_controller));
+                                  nullptr, table_cache.get(),
+                                  &write_buffer_manager, &write_controller));
 
     wal_manager.reset(new WalManager(db_options, env_options));