+++ HISTORY.md
@@ -6,6 +6,7 @@
 ### New Features
 * Measure estimated number of reads per file. The information can be accessed through DB::GetColumnFamilyMetaData or "rocksdb.sstables" DB property.
 * RateLimiter support for throttling background reads, or throttling the sum of background reads and writes. This can give more predictable I/O usage when compaction reads more data than it writes, e.g., due to lots of deletions.
+* In leveled compaction, if an SST file is read too many times, and we don't have anything else to compact, the file will be compacted to the next level. Now it is only applied for Get().
 
 ## 5.6.0 (06/06/2017)
 ### Public API Change